Add length limit to intra-word diffs to prevent exception from being thrown
Summary:
This diff limits the length of string that we will use intra-word diffs as strings longer than 255 will cause the \levenshtein function to throw an exception.

## Context

When writing tests, people sometimes compared long error strings (length > 255). When they called `expect(string)->toEqual(string2)` on long functions and they weren't the same, the tests called on `ColoredUnifiedDiff` to return a colored difference of the two strings.
The old ColoredUnifiedDiff included a call to the global PHP \levenshtein distance function. However, when comparing two long strings, this function throws an exception which fails the test with a cryptic `ErrorException: levenshtein(): Argument string(s) too long`. By limiting the length of the string we can call the levenshtein function with, we avoid an exception.

Reviewed By: fredemmott

Differential Revision: D20603044

fbshipit-source-id: 19abf15087be3337075bb672f7ef4332394b7efb